Refactor: Merge determine_optimal_transform and lib into transform module in bc1
