+++ projects/api/dxt-lossless-transform-bc1-api/src/transform/mod.rs
@@ -9,9 +9,7 @@ use crate::error::Bc1Error;
 use dxt_lossless_transform_api_common::estimate::SizeEstimationOperations;
 use dxt_lossless_transform_bc1::{
     Bc1TransformSettings,
-    determine_optimal_transform::{
-        Bc1EstimateOptions, DetermineBestTransformError, transform_with_best_options,
-    },
+    transform::{Bc1EstimateOptions, DetermineBestTransformError, transform_with_best_options},
 };
 
 /// Transform BC1 data using automatically determined optimal settings.
